Withdraw envelope (C-quorum): gate cPoC on each verifier's local oracle.
Request-leg originator labels are unsigned, so counting them as independent
host votes was unsound; keep Q only for turn reachability and floor raises.
